Move kill() and is_running() to NetConf
The only consumer of kill() and is_running() is NetConf
